+++ .gitignore
@@ -56,3 +56,5 @@ next-env.d.ts
 .zed
 .tempor
 .agents
+
+.patches
