+++ src/modules/application/components/layouts/layoutDao/layoutDao.tsx
@@ -4,10 +4,10 @@ import {
     QueryClient,
 } from '@tanstack/react-query';
 import type { ReactNode } from 'react';
-import { daoOverridesOptions } from '@/modules/explore/api/cmsService';
 import { LockToVoteLockOnboardingWatcher } from '@/plugins/lockToVotePlugin/components/lockToVoteLockOnboardingWatcher';
 import { TokenDelegationOnboardingWatcher } from '@/plugins/tokenPlugin/components/tokenDelegationOnboardingWatcher';
 import { TokenLockAndWrapOnboardingWatcher } from '@/plugins/tokenPlugin/components/tokenLockAndWrapOnboardingWatcher';
+import { daoOverridesOptions } from '@/shared/api/cmsService';
 import { daoOptions, type IDao } from '@/shared/api/daoService';
 import { Page } from '@/shared/components/page';
 import type { IDaoPageParams } from '@/shared/types';
