+++ src/modules/application/components/layouts/layoutRoot/layoutRoot.tsx
@@ -7,8 +7,8 @@ import { initActionViewRegistry } from '@/actions';
 import { initPluginRegistry } from '@/initPluginRegistry';
 import { wagmiConfig } from '@/modules/application/constants/wagmi';
 import { fetchInterceptorUtils } from '@/modules/application/utils/fetchInterceptorUtils';
-import { sanctionedAddressesOptions } from '@/modules/explore/api/cmsService';
-import { whitelistedAddressesOptions } from '@/modules/explore/api/cmsService/queries/useWhitelistedAddresses';
+import { sanctionedAddressesOptions } from '@/shared/api/cmsService';
+import { whitelistedAddressesOptions } from '@/shared/api/cmsService/queries/useWhitelistedAddresses';
 import { translations } from '@/shared/constants/translations';
 import { featureFlags } from '@/shared/featureFlags';
 import { DebugPanelLazy } from '../../debugPanel/lazyDebugPanel';
