+++ src/modules/explore/pages/exploreDaosPage/exploreDaosPageClient.tsx
@@ -5,6 +5,7 @@ import classNames from 'classnames';
 import Image from 'next/image';
 import { Suspense } from 'react';
 import { CreateDaoDialogId } from '@/modules/createDao/constants/createDaoDialogId';
+import { useFeaturedDaos } from '@/shared/api/cmsService';
 import { Carousel } from '@/shared/components/carousel';
 import { Container } from '@/shared/components/container';
 import { useDialogContext } from '@/shared/components/dialogProvider';
@@ -13,7 +14,6 @@ import doItYourselfIcon from '../../../../assets/images/doItYourselfIcon.svg';
 import enterpriseServiceIcon from '../../../../assets/images/enterpriseServiceIcon.svg';
 import NetBackground from '../../../../assets/images/net_bg.svg';
 import noCodeSetupIcon from '../../../../assets/images/noCodeSetup.svg';
-import { useFeaturedDaos } from '../../api/cmsService';
 import type { IGetDaoListParams } from '../../api/daoExplorerService';
 import { CtaCard } from '../../components/ctaCard';
 import { DaoCarouselCard } from '../../components/daoCarouselCard';
