+++ src/modules/governance/pages/daoProposalsPage/daoProposalsPage.tsx
@@ -1,5 +1,5 @@
 import { QueryClient } from '@tanstack/react-query';
-import { daoOverridesOptions } from '@/modules/explore/api/cmsService';
+import { daoOverridesOptions } from '@/shared/api/cmsService';
 import { daoOptions } from '@/shared/api/daoService';
 import { Page } from '@/shared/components/page';
 import { RedirectToUrl } from '@/shared/components/redirectToUrl';
