+++ src/shared/api/cmsService/domain/sanctionedAddressesResult.ts
